fix: bound BackendTrafficPolicy rateLimit requests to uint32 max (#8798)
* api: bound BackendTrafficPolicy rateLimit requests to uint32 max

The RLS config proto (RateLimitPolicy.requests_per_unit) and Envoy's
TokenBucket (max_tokens / tokens_per_fill) are uint32 on the wire.
RateLimitValue.Requests was unbounded in the CRD and narrowed with an
unchecked uint32(...) cast in the xDS translator, so values above
4294967295 were silently truncated modulo 2^32 (e.g. 30000000000 -> 4230196224).

Add kubebuilder Minimum=1 / Maximum=4294967295 markers so such values are
rejected at admission with a clear schema error, and regenerate the CRDs
and helm-template snapshots. The existing uint32(...) narrowings are now
safe by construction.

Includes CEL validation coverage for the boundary, the above-max case
(using the 30B repro value), the zero case, and the local rate-limit path.

Fixes #8797
Refs envoyproxy/ai-gateway#2012

Made-with: Cursor

* api: align rate-limit requests with uint32 limits

Made-with: Cursor
Made-with: Cursor

* test: exercise uint32 max in rate-limit testdata

Bumps one rule in backendtrafficpolicy-with-ratelimit (gatewayapi) and
multiple-rules (xds/translator RLS config) to 4294967295 to confirm the
uint32 boundary renders end-to-end as requests_per_unit: 4294967295
without wraparound. Addresses review feedback on #8798.


* api: trim RateLimitValue.Requests docstring


* xds/translator: drop redundant uint32 casts on RateLimitValue.Requests


* fit lint and minor refactoring


---------

+++ internal/xds/translator/ratelimit.go
@@ -817,7 +817,7 @@ func buildRateLimitServiceDescriptors(route *ir.HTTPRoute) []*rlsconfv3.RateLimi
 
 	for rIdx, rule := range global.Rules {
 		rateLimitPolicy := &rlsconfv3.RateLimitPolicy{
-			RequestsPerUnit: uint32(rule.Limit.Requests),
+			RequestsPerUnit: rule.Limit.Requests,
 			Unit: rlsconfv3.RateLimitUnit(
 				rlsconfv3.RateLimitUnit_value[strings.ToUpper(string(rule.Limit.Unit))]),
 		}
